+++ src/Application/Query/PromotionCode/Batch/PromotionCodeGroupList/GetListView.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Proximum\Vimeet\Application\Query\PromotionCode\Batch\PromotionCodeGroupList;
+
+use Proximum\Vimeet\Domain\Model\Event;
+
+class GetListView
+{
+    /** @var Event */
+    public $event;
+
+    public function __construct(Event $event)
+    {
+        $this->event = $event;
+    }
+}